Removed a compiler warning
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/535



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@2135 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/unicode/unicode.h
@@ -1213,7 +1213,7 @@ Unicode::digit(UChar32 ch, int8_t radix) {
     } else {
         return -1; // invalid radix
     }
-    return (value<radix) ? value : -1;
+    return (uint8_t)((value<radix) ? value : (uint8_t)(-1));
 }
 
 inline UChar32
